simplify schema code
